Generate onig config.h on non win32 OSes
Trying to compile edbee-lib on non-64 bit unix like OSes can lead to issues due
to compile time checks in the onig configuration, since the checked in config.h
was created for 64 bit OSes. This commit will run the configure script for each
compile run to make sure the header is adapted for the current OS type.

It should be considered to remove the config.h file from the repo and put it on
gitignore.