Update: bring QsLog up to date
This commit replaces the code originally used in pretty much the first
commit with the current version of the QsLog code from:
https://bitbucket.org/codeimproved/qslog as of commit: 46b643d5bcbc .

This was done because the original code is showing signs of its age with
current compilers and throwing up warnings.

The option to include file and line numbers has changed in name slightly
and the qmake and cmake project files from the updated source code have
been amended to have that option enabled by default as the original code
had.

Some superfluous "example" code has been pruned as a sub-directory in the
replacement code used.

It was also necessary to modify the logging.h header in the edbee/util
superproject directory so that it used a non-system #include for the
QsLog header.

I have updated the Qt include project file so that it lists some more files
in the Qt Creator (the Qt IDE) relating to the main edbee-lib superproject.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>